backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["TobiFank <tobias7402@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
pillow = "^11.0.0"
openai = "^1.58.1"
transformers = "^4.47.1"
torch = "^2.5.1"
tqdm = "^4.67.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
